+++ test/error_msg_tests.js
@@ -3,7 +3,7 @@ import errorMsgs from '../src/errorMsgs.js'
 
 var assertionCnt = 0;
 
-function checkError(msgList, line, offset, msg) {
+export function checkError(msgList, line, offset, msg) {
     assertionCnt++;
     for( const i in msgList ) {
         if( msgList[i].from.line == line && msgList[i].from.depth == offset && msgList[i].message == msg ) {
@@ -27,7 +27,7 @@ export function checkNoError(msgList, msgPrefix) {
 
 var output;
 
-export default function error_msg_tests() {
+export function error_msg_tests() {
 
     output =toErrors(`dept
     id
@@ -38,7 +38,7 @@ export default function error_msg_tests() {
     name vc-200
     name vc0
     `).getErrors();
-    checkError(output, 1, 4+4+2+1, errorMsgs.messages.invalidDatatype);
+    checkError(output, 1, 4 + 4 + 2 + 1, errorMsgs.messages.invalidDatatype);
     checkError(output, 2, 4+4+1, errorMsgs.messages.invalidDatatype);
     checkNoError(output, errorMsgs.messages.misalignedAttribute);
 
@@ -106,15 +106,31 @@ team_statuses
     name /fk undefined
     `).getErrors();
     checkError(output, 3, 4+4+1+3+1, errorMsgs.messages.undefinedObject+'undefined');
+
+    output = new quicksql(`emp
+    ename
+    deptno /fk dept
+  
+dept 
+   dname`).getErrors();
+    checkError(output, 5, 3, errorMsgs.messages.misalignedAttribute+"4");
+  
+    output = new quicksql(`emp /fk
+    ename  /audit
+    `).getErrors();
+    checkError(output, 0, 5, errorMsgs.messages.tableDirectiveTypo);
+    checkError(output, 1, 4+5+1+1+1, errorMsgs.messages.columnDirectiveTypo);
     
+    console.log(assertionCnt);
+
+    const minimalTestCnt = 10;
+    if( assertionCnt < minimalTestCnt ) {
+        console.error("assertionCnt < "+minimalTestCnt);
+        throw new Error('Test failed');
+    } 
+ 
 }
 
-error_msg_tests();
 
-console.log(assertionCnt);
 
-const minimalTestCnt = 10;
-if( assertionCnt < minimalTestCnt ) {
-    console.error("assertionCnt < "+minimalTestCnt);
-    throw new Error('Test failed');
-} 
+